+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 ### Added
 - feat: Add sections headers
-- feat: Display alerts
+- feat: Display alerts (#110)
 - style(lint): Add eslint plugin to avoid the usage of `only` in tests
 ### Changed
 - refactor: Move render helpers to a helpers file