+++ CHANGELOG.md
@@ -6,7 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 ### Added
+- feat: Add sections headers
+- feat: Display alerts (#110)
+- style(lint): Add eslint plugin to avoid the usage of `only` in tests
 ### Changed
+- refactor: Move render helpers to a helpers file
+- refactor: Remove unused inquirerCli option
+- test(e2e): Move support elements to support folder
+- docs(readme): Add main features chapter
 ### Fixed
 ### Removed